pdf-accounting-importer/backend/requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
pillow
numpy
paddleocr
paddlepaddle
opencv-contrib-python
